core(lantern): rename lib/dependency-graph to lib/lantern
