+++ core/computed/metrics/lantern-first-meaningful-paint.js
@@ -9,7 +9,7 @@ import {LanternMetric} from './lantern-metric.js';
 import {LighthouseError} from '../../lib/lh-error.js';
 import {LanternFirstContentfulPaint} from './lantern-first-contentful-paint.js';
 
-/** @typedef {import('../../lib/dependency-graph/base-node.js').Node<LH.Artifacts.NetworkRequest>} Node */
+/** @typedef {import('../../lib/lantern/base-node.js').Node<LH.Artifacts.NetworkRequest>} Node */
 
 class LanternFirstMeaningfulPaint extends LanternMetric {
   /**
