+++ core/lib/asset-saver.js
@@ -11,7 +11,7 @@ import url from 'url';
 
 import log from 'lighthouse-logger';
 
-import {Simulator} from './dependency-graph/simulator/simulator.js';
+import {Simulator} from './lantern/simulator/simulator.js';
 import lanternTraceSaver from './lantern-trace-saver.js';
 import {MetricTraceEvents} from './traces/metric-trace-events.js';
 import {NetworkAnalysis} from '../computed/network-analysis.js';
