+++ core/lib/lantern/base-node.js
