+++ core/lib/lantern/cpu-node.js
