+++ core/lib/lantern/lantern.js
