+++ core/lib/lantern/network-node.js
