+++ core/lib/lantern/simulator/connection-pool.js
