+++ core/lib/lantern/simulator/dns-cache.js
