+++ core/lib/lantern/simulator/network-analyzer.js
