+++ core/lib/lantern/simulator/simulator-timing-map.js
