+++ core/lib/lantern/simulator/simulator.js
