+++ core/lib/lantern/simulator/tcp-connection.js
