+++ core/scripts/lantern/run-once.js
@@ -9,7 +9,7 @@ import fs from 'fs';
 import path from 'path';
 
 import PredictivePerf from '../../audits/predictive-perf.js';
-import {Simulator} from '../../lib/dependency-graph/simulator/simulator.js';
+import {Simulator} from '../../lib/lantern/simulator/simulator.js';
 import traceSaver from '../../lib/lantern-trace-saver.js';
 import {LH_ROOT} from '../../../shared/root.js';
 import {readJson} from '../../test/test-utils.js';
