+++ core/test/audits/byte-efficiency/render-blocking-resources-test.js
@@ -8,9 +8,9 @@ import assert from 'assert/strict';
 
 import RenderBlockingResourcesAudit from '../../../audits/byte-efficiency/render-blocking-resources.js'; // eslint-disable-line max-len
 import * as constants from '../../../config/constants.js';
-import {NetworkNode} from '../../../lib/dependency-graph/network-node.js';
-import {CPUNode} from '../../../lib/dependency-graph/cpu-node.js';
-import {Simulator} from '../../../lib/dependency-graph/simulator/simulator.js';
+import {NetworkNode} from '../../../lib/lantern/network-node.js';
+import {CPUNode} from '../../../lib/lantern/cpu-node.js';
+import {Simulator} from '../../../lib/lantern/simulator/simulator.js';
 import {NetworkRequest} from '../../../lib/network-request.js';
 import {getURLArtifactFromDevtoolsLog, readJson} from '../../test-utils.js';
