+++ tsconfig.json
@@ -65,14 +65,14 @@
     "core/test/index-test.js",
     "core/test/lib/arbitrary-equality-map-test.js",
     "core/test/lib/asset-saver-test.js",
-    "core/test/lib/dependency-graph/base-node-test.js",
-    "core/test/lib/dependency-graph/simulator/connection-pool-test.js",
-    "core/test/lib/dependency-graph/simulator/dns-cache-test.js",
-    "core/test/lib/dependency-graph/simulator/network-analyzer-test.js",
-    "core/test/lib/dependency-graph/simulator/simulator-test.js",
     "core/test/lib/emulation-test.js",
     "core/test/lib/i18n/i18n-test.js",
     "core/test/lib/icons-test.js",
+    "core/test/lib/lantern/base-node-test.js",
+    "core/test/lib/lantern/simulator/connection-pool-test.js",
+    "core/test/lib/lantern/simulator/dns-cache-test.js",
+    "core/test/lib/lantern/simulator/network-analyzer-test.js",
+    "core/test/lib/lantern/simulator/simulator-test.js",
     "core/test/lib/lh-element-test.js",
     "core/test/lib/lighthouse-compatibility-test.js",
     "core/test/lib/manifest-parser-test.js",
